pyproject.toml
[project]
name = "gemi9"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["google-adk==1.16.0"]
